remove wimpy dependency
